requirements.txt
fastapi
uvicorn
pillow
python-magic
tensorflow
keras
numpy
python-multipart